Refactor code for @link to address some long-standing issues/todos
