+++ apollo-federation/src/link/mod.rs
@@ -1,26 +1,17 @@
-use std::collections::HashMap;
 use std::fmt;
 use std::ops::Range;
 use std::str;
 use std::sync::Arc;
 
-use apollo_compiler::InvalidNameError;
 use apollo_compiler::Name;
 use apollo_compiler::Node;
 use apollo_compiler::Schema;
 use apollo_compiler::ast::Directive;
 use apollo_compiler::ast::Value;
-use apollo_compiler::collections::IndexMap;
-use apollo_compiler::name;
 use apollo_compiler::parser::LineColumn;
 use apollo_compiler::schema::Component;
-use thiserror::Error;
 
 use crate::error::FederationError;
-use crate::error::SingleFederationError;
-use crate::link::link_spec_definition::CORE_VERSIONS;
-use crate::link::link_spec_definition::LINK_VERSIONS;
-use crate::link::link_spec_definition::LinkSpecDefinition;
 use crate::link::spec::Identity;
 use crate::link::spec::Url;
 
@@ -29,45 +20,18 @@ pub(crate) mod authenticated_spec_definition;
 pub(crate) mod cache_tag_spec_definition;
 pub(crate) mod context_spec_definition;
 pub mod cost_spec_definition;
-pub mod database;
 pub(crate) mod federation_spec_definition;
 pub(crate) mod graphql_definition;
 pub(crate) mod inaccessible_spec_definition;
 pub(crate) mod join_spec_definition;
 pub(crate) mod link_spec_definition;
+pub mod metadata;
 pub(crate) mod policy_spec_definition;
 pub(crate) mod requires_scopes_spec_definition;
 pub mod spec;
 pub(crate) mod spec_definition;
+pub(crate) mod spec_registry;
 pub(crate) mod tag_spec_definition;
-
-pub const DEFAULT_LINK_NAME: Name = name!("link");
-pub const DEFAULT_IMPORT_SCALAR_NAME: Name = name!("Import");
-pub const DEFAULT_PURPOSE_ENUM_NAME: Name = name!("Purpose");
-pub(crate) const IMPORT_AS_ARGUMENT: Name = name!("as");
-pub(crate) const IMPORT_NAME_ARGUMENT: Name = name!("name");
-
-// TODO: we should provide proper "diagnostic" here, linking to ast, accumulating more than one
-// error and whatnot.
-#[derive(Error, Debug, PartialEq)]
-pub enum LinkError {
-    #[error(transparent)]
-    InvalidName(#[from] InvalidNameError),
-    #[error("Invalid use of @link in schema: {0}")]
-    BootstrapError(String),
-    #[error("Unknown import: {0}")]
-    InvalidImport(String),
-}
-
-// TODO: Replace LinkError usages with FederationError.
-impl From<LinkError> for FederationError {
-    fn from(value: LinkError) -> Self {
-        SingleFederationError::InvalidLinkDirectiveUsage {
-            message: value.to_string(),
-        }
-        .into()
-    }
-}
 
 #[derive(Clone, Copy, Eq, PartialEq, Debug)]
 pub enum Purpose {
@@ -75,46 +39,9 @@ pub enum Purpose {
     EXECUTION,
 }
 
-impl Purpose {
-    pub fn from_value(value: &Value) -> Result<Purpose, LinkError> {
-        value
-            .as_enum()
-            .ok_or_else(|| {
-                LinkError::BootstrapError("invalid `purpose` value, should be an enum".to_string())
-            })
-            .and_then(|value| value.parse())
-    }
-}
-
-impl str::FromStr for Purpose {
-    type Err = LinkError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "SECURITY" => Ok(Purpose::SECURITY),
-            "EXECUTION" => Ok(Purpose::EXECUTION),
-            _ => Err(LinkError::BootstrapError(format!(
-                "invalid/unrecognized `purpose` value '{s}'"
-            ))),
-        }
-    }
-}
-
 impl fmt::Display for Purpose {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Purpose::SECURITY => f.write_str("SECURITY"),
-            Purpose::EXECUTION => f.write_str("EXECUTION"),
-        }
-    }
-}
-
-impl From<&Purpose> for Name {
-    fn from(value: &Purpose) -> Self {
-        match value {
-            Purpose::SECURITY => name!("SECURITY"),
-            Purpose::EXECUTION => name!("EXECUTION"),
-        }
+        Value::from(self).fmt(f)
     }
 }
 
@@ -134,86 +61,8 @@ pub struct Import {
 }
 
 impl Import {
-    pub fn from_value(value: &Value) -> Result<Import, LinkError> {
-        // TODO: it could be nice to include the broken value in the error messages of this method
-        // (especially since @link(import:) is a list), but `Value` does not implement `Display`
-        // currently, so a bit annoying.
-        match value {
-            Value::String(str) => {
-                if let Some(directive_name) = str.strip_prefix('@') {
-                    Ok(Import {
-                        element: Name::new(directive_name)?,
-                        is_directive: true,
-                        alias: None,
-                    })
-                } else {
-                    Ok(Import {
-                        element: Name::new(str)?,
-                        is_directive: false,
-                        alias: None,
-                    })
-                }
-            }
-            Value::Object(fields) => {
-                let mut name: Option<&str> = None;
-                let mut alias: Option<&str> = None;
-                for (k, v) in fields {
-                    match k.as_str() {
-                        "name" => {
-                            name = Some(v.as_str().ok_or_else(|| {
-                                LinkError::BootstrapError(format!(r#"in "{}", invalid value for `name` field in @link(import:) argument: must be a string"#, value.serialize().no_indent()))
-                            })?)
-                        },
-                        "as" => {
-                            alias = Some(v.as_str().ok_or_else(|| {
-                                LinkError::BootstrapError(format!(r#"in "{}", invalid value for `as` field in @link(import:) argument: must be a string"#, value.serialize().no_indent()))
-                            })?)
-                        },
-                        _ => Err(LinkError::BootstrapError(format!(r#"in "{}", unknown field `{k}` in @link(import:) argument"#, value.serialize().no_indent())))?
-                    }
-                }
-                let Some(element) = name else {
-                    return Err(LinkError::BootstrapError(format!(
-                        r#"in "{}", invalid entry in @link(import:) argument, missing mandatory `name` field"#,
-                        value.serialize().no_indent()
-                    )));
-                };
-                if let Some(directive_name) = element.strip_prefix('@') {
-                    if let Some(alias_str) = alias.as_ref() {
-                        let Some(alias_str) = alias_str.strip_prefix('@') else {
-                            return Err(LinkError::BootstrapError(format!(
-                                r#"in "{}", invalid alias '{alias_str}' for import name '{element}': should start with '@' since the imported name does"#,
-                                value.serialize().no_indent()
-                            )));
-                        };
-                        alias = Some(alias_str);
-                    }
-                    Ok(Import {
-                        element: Name::new(directive_name)?,
-                        is_directive: true,
-                        alias: alias.map(Name::new).transpose()?,
-                    })
-                } else {
-                    if let Some(alias) = &alias
-                        && alias.starts_with('@')
-                    {
-                        return Err(LinkError::BootstrapError(format!(
-                            r#"in "{}", invalid alias '{alias}' for import name '{element}': should not start with '@' (or, if {element} is a directive, then the name should start with '@')"#,
-                            value.serialize().no_indent()
-                        )));
-                    }
-                    Ok(Import {
-                        element: Name::new(element)?,
-                        is_directive: false,
-                        alias: alias.map(Name::new).transpose()?,
-                    })
-                }
-            }
-            _ => Err(LinkError::BootstrapError(format!(
-                r#"in "{}", invalid sub-value for @link(import:) argument: values should be either strings or input object values of the form {{ name: "<importedElement>", as: "<alias>" }}."#,
-                value.serialize().no_indent()
-            ))),
-        }
+    pub fn imported_name(&self) -> &Name {
+        self.alias.as_ref().unwrap_or(&self.element)
     }
 
     pub fn element_display_name(&self) -> impl fmt::Display {
@@ -221,10 +70,6 @@ impl Import {
             name: &self.element,
             is_directive: self.is_directive,
         }
-    }
-
-    pub fn imported_name(&self) -> &Name {
-        self.alias.as_ref().unwrap_or(&self.element)
     }
 
     pub fn imported_display_name(&self) -> impl fmt::Display {
@@ -252,44 +97,7 @@ impl fmt::Display for DisplayName<'_> {
 
 impl fmt::Display for Import {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if self.alias.is_some() {
-            write!(
-                f,
-                r#"{{ name: "{}", as: "{}" }}"#,
-                self.element_display_name(),
-                self.imported_display_name()
-            )
-        } else {
-            write!(f, r#""{}""#, self.imported_display_name())
-        }
-    }
-}
-
-#[allow(clippy::from_over_into)]
-impl Into<Value> for Import {
-    fn into(self) -> Value {
-        let element_string = if self.is_directive {
-            format!("@{}", self.element)
-        } else {
-            self.element.to_string()
-        };
-
-        if let Some(alias) = self.alias {
-            let alias_string = if self.is_directive {
-                format!("@{}", alias)
-            } else {
-                alias.to_string()
-            };
-            Value::Object(vec![
-                (
-                    IMPORT_NAME_ARGUMENT,
-                    Node::new(Value::String(element_string)),
-                ),
-                (IMPORT_AS_ARGUMENT, Node::new(Value::String(alias_string))),
-            ])
-        } else {
-            Value::String(element_string)
-        }
+        Value::from(self).fmt(f)
     }
 }
 
@@ -303,6 +111,15 @@ pub struct Link {
 }
 
 impl Link {
+    pub fn from_directive_application(
+        directive: &Node<Directive>,
+        schema: &Schema,
+    ) -> Result<Link, FederationError> {
+        let mut link = Link::try_from(directive.as_ref())?;
+        link.line_column_range = directive.line_column_range(&schema.sources);
+        Ok(link)
+    }
+
     pub fn spec_name_in_schema(&self) -> &Name {
         self.spec_alias.as_ref().unwrap_or(&self.url.identity.name)
     }
@@ -351,71 +168,6 @@ impl Link {
         }
     }
 
-    pub fn from_directive_application(
-        directive: &Node<Directive>,
-        schema: &Schema,
-    ) -> Result<Link, LinkError> {
-        let (url, is_link) = if let Some(value) = directive.specified_argument_by_name("url") {
-            (value, true)
-        } else if let Some(value) = directive.specified_argument_by_name("feature") {
-            // XXX(@goto-bus-stop): @core compatibility is primarily to support old tests--should be
-            // removed when those are updated.
-            (value, false)
-        } else {
-            return Err(LinkError::BootstrapError(
-                "the `url` argument for @link is mandatory".to_string(),
-            ));
-        };
-
-        let (directive_name, arg_name) = if is_link {
-            ("link", "url")
-        } else {
-            ("core", "feature")
-        };
-
-        let url = url.as_str().ok_or_else(|| {
-            LinkError::BootstrapError(format!(
-                "the `{arg_name}` argument for @{directive_name} must be a String"
-            ))
-        })?;
-        let url: Url = url.parse::<Url>().map_err(|e| {
-            LinkError::BootstrapError(format!("invalid `{arg_name}` argument (reason: {e})"))
-        })?;
-
-        let spec_alias = directive
-            .specified_argument_by_name("as")
-            .and_then(|arg| arg.as_str())
-            .map(Name::new)
-            .transpose()?;
-        let purpose = if let Some(value) = directive.specified_argument_by_name("for") {
-            Some(Purpose::from_value(value)?)
-        } else {
-            None
-        };
-
-        let imports = if is_link {
-            directive
-                .specified_argument_by_name("import")
-                .and_then(|arg| arg.as_list())
-                .unwrap_or(&[])
-                .iter()
-                .map(|value| Ok(Arc::new(Import::from_value(value)?)))
-                .collect::<Result<Vec<Arc<Import>>, LinkError>>()?
-        } else {
-            Default::default()
-        };
-
-        let line_column_range = directive.line_column_range(&schema.sources);
-
-        Ok(Link {
-            url,
-            spec_alias,
-            imports,
-            purpose,
-            line_column_range,
-        })
-    }
-
     pub fn for_identity<'schema>(
         schema: &'schema Schema,
         identity: &Identity,
@@ -445,161 +197,6 @@ impl Link {
 
 impl fmt::Display for Link {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let imported_types: Vec<String> = self
-            .imports
-            .iter()
-            .map(|import| import.to_string())
-            .collect::<Vec<String>>();
-        let imports = if imported_types.is_empty() {
-            "".to_string()
-        } else {
-            format!(r#", import: [{}]"#, imported_types.join(", "))
-        };
-        let alias = self
-            .spec_alias
-            .as_ref()
-            .map(|a| format!(r#", as: "{a}""#))
-            .unwrap_or("".to_string());
-        let purpose = self
-            .purpose
-            .as_ref()
-            .map(|p| format!(r#", for: {p}"#))
-            .unwrap_or("".to_string());
-        write!(f, r#"@link(url: "{}"{alias}{imports}{purpose})"#, self.url)
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct LinkedElement {
-    pub link: Arc<Link>,
-    pub import: Option<Arc<Import>>,
-    pub name: Name,
-    pub name_in_spec: Name,
-}
-
-#[derive(Clone, Eq, PartialEq, Debug)]
-pub struct LinksMetadata {
-    pub(crate) links: Vec<Arc<Link>>,
-    pub(crate) by_identity: IndexMap<Identity, Arc<Link>>,
-    pub(crate) by_name_in_schema: IndexMap<Name, Arc<Link>>,
-    pub(crate) types_by_imported_name: IndexMap<Name, (Arc<Link>, Arc<Import>)>,
-    pub(crate) directives_by_imported_name: IndexMap<Name, (Arc<Link>, Arc<Import>)>,
-}
-
-impl LinksMetadata {
-    // PORT_NOTE: Call this as a replacement for `CoreFeatures.coreItself` from JS.
-    pub(crate) fn link_spec_definition(
-        &self,
-    ) -> Result<&'static LinkSpecDefinition, FederationError> {
-        if let Some(link_link) = self.for_identity(&Identity::link_identity()) {
-            LINK_VERSIONS.find(&link_link.url.version).ok_or_else(|| {
-                SingleFederationError::Internal {
-                    message: format!("Unexpected link spec version {}", link_link.url.version),
-                }
-                .into()
-            })
-        } else if let Some(core_link) = self.for_identity(&Identity::core_identity()) {
-            CORE_VERSIONS.find(&core_link.url.version).ok_or_else(|| {
-                SingleFederationError::Internal {
-                    message: format!("Unexpected core spec version {}", core_link.url.version),
-                }
-                .into()
-            })
-        } else {
-            Err(SingleFederationError::Internal {
-                message: "Unexpectedly could not find core/link spec".to_owned(),
-            }
-            .into())
-        }
-    }
-
-    pub fn all_links(&self) -> &[Arc<Link>] {
-        self.links.as_ref()
-    }
-
-    pub fn for_identity(&self, identity: &Identity) -> Option<Arc<Link>> {
-        self.by_identity.get(identity).cloned()
-    }
-
-    pub fn source_link_of_type(&self, type_name: &Name) -> Option<LinkedElement> {
-        // For types, it's either an imported name or it must be fully qualified
-        if let Some((link, import)) = self.types_by_imported_name.get(type_name) {
-            return Some(LinkedElement {
-                link: Arc::clone(link),
-                import: Some(Arc::clone(import)),
-                name: type_name.clone(),
-                name_in_spec: import.element.clone(),
-            });
-        }
-
-        type_name
-            .split_once("__")
-            .and_then(|(spec_name, name_in_spec)| {
-                let Ok(name_in_spec) = Name::new(name_in_spec) else {
-                    return None;
-                };
-                self.by_name_in_schema
-                    .get(spec_name)
-                    .map(|link| LinkedElement {
-                        link: Arc::clone(link),
-                        import: None,
-                        name: type_name.clone(),
-                        name_in_spec,
-                    })
-            })
-    }
-
-    pub fn source_link_of_directive(&self, directive_name: &Name) -> Option<LinkedElement> {
-        // For directives, it can be either:
-        //   1. be an imported name,
-        //   2. be the "imported" name of a linked spec (special case of a directive named like the
-        //      spec),
-        //   3. or it must be fully qualified.
-        if let Some((link, import)) = self.directives_by_imported_name.get(directive_name) {
-            return Some(LinkedElement {
-                link: Arc::clone(link),
-                import: Some(Arc::clone(import)),
-                name: directive_name.clone(),
-                name_in_spec: import.element.clone(),
-            });
-        }
-
-        if let Some(link) = self.by_name_in_schema.get(directive_name) {
-            return Some(LinkedElement {
-                link: Arc::clone(link),
-                import: None,
-                name: directive_name.clone(),
-                name_in_spec: link.url.identity.name.clone(),
-            });
-        }
-
-        directive_name
-            .split_once("__")
-            .and_then(|(spec_name, name_in_spec)| {
-                let Ok(name_in_spec) = Name::new(name_in_spec) else {
-                    return None;
-                };
-                self.by_name_in_schema
-                    .get(spec_name)
-                    .map(|link| LinkedElement {
-                        link: Arc::clone(link),
-                        import: None,
-                        name: directive_name.clone(),
-                        name_in_spec,
-                    })
-            })
-    }
-
-    pub(crate) fn import_to_feature_url_map(&self) -> HashMap<String, Url> {
-        let directive_entries = self
-            .directives_by_imported_name
-            .iter()
-            .map(|(name, (link, _))| (name.to_string(), link.url.clone()));
-        let type_entries = self
-            .types_by_imported_name
-            .iter()
-            .map(|(name, (link, _))| (name.to_string(), link.url.clone()));
-
-        directive_entries.chain(type_entries).collect()
+        Directive::from(self).fmt(f)
     }
 }
