+++ apollo-federation/src/link/spec_definition.rs
@@ -1,9 +1,6 @@
 use std::collections::BTreeMap;
-use std::collections::BTreeSet;
-use std::collections::HashMap;
 use std::collections::btree_map::Keys;
 use std::sync::Arc;
-use std::sync::LazyLock;
 
 use apollo_compiler::Name;
 use apollo_compiler::Node;
@@ -12,30 +9,16 @@ use apollo_compiler::schema::DirectiveDefinition;
 use apollo_compiler::schema::ExtendedType;
 use itertools::Itertools;
 
-use crate::AUTHENTICATED_VERSIONS;
-use crate::CACHE_TAG_VERSIONS;
-use crate::CONTEXT_VERSIONS;
-use crate::COST_VERSIONS;
-use crate::INACCESSIBLE_VERSIONS;
-use crate::POLICY_VERSIONS;
-use crate::REQUIRES_SCOPES_VERSIONS;
-use crate::TAG_VERSIONS;
-use crate::connectors::spec::CONNECT_VERSIONS;
 use crate::ensure;
 use crate::error::FederationError;
 use crate::error::MultipleFederationErrors;
 use crate::error::SingleFederationError;
-use crate::link::Import;
 use crate::link::Link;
 use crate::link::Purpose;
-use crate::link::federation_spec_definition::FEDERATION_VERSIONS;
-use crate::link::join_spec_definition::JOIN_VERSIONS;
 use crate::link::spec::Identity;
 use crate::link::spec::Url;
 use crate::link::spec::Version;
 use crate::schema::FederationSchema;
-use crate::schema::type_and_directive_specification::DirectiveCompositionSpecification;
-use crate::schema::type_and_directive_specification::DirectiveSpecification;
 use crate::schema::type_and_directive_specification::TypeAndDirectiveSpecification;
 
 pub(crate) trait SpecDefinition {
@@ -296,69 +279,3 @@ impl<T: SpecDefinition> SpecDefinitions<T> {
             .map(|spec| spec as &dyn SpecDefinition)
     }
 }
-
-pub(crate) struct SpecRegistry {
-    definitions_by_url: HashMap<Url, &'static (dyn SpecDefinition + Sync)>,
-    available_versions_by_identity: HashMap<Identity, BTreeSet<Version>>,
-}
-
-impl SpecRegistry {
-    fn new() -> Self {
-        Self {
-            definitions_by_url: HashMap::new(),
-            available_versions_by_identity: HashMap::new(),
-        }
-    }
-
-    fn extend<T: SpecDefinition + Sync>(&mut self, definitions: &'static SpecDefinitions<T>) {
-        for (v, spec) in definitions.iter() {
-            self.definitions_by_url.insert(spec.url().clone(), spec);
-            self.available_versions_by_identity
-                .entry(spec.url().identity.clone())
-                .or_default()
-                .insert(v.clone());
-        }
-    }
-
-    pub(crate) fn get_definition(&self, url: &Url) -> Option<&&(dyn SpecDefinition + Sync)> {
-        self.definitions_by_url.get(url)
-    }
-
-    pub(crate) fn get_versions(&self, identity: &Identity) -> Option<&BTreeSet<Version>> {
-        self.available_versions_by_identity.get(identity)
-    }
-
-    /// Generates the composition spec for an imported directive. Currently, this generates the
-    /// entire spec, then loops over available directive specifications and clones the applicable
-    /// directive. An alternative would be to mark everything as `Sync` and store them on the
-    /// individual feature specs, but we have omitted this for now due to a non-trivial (~10%)
-    /// increase in heap usage that affects query planning.
-    pub(crate) fn get_composition_spec(
-        &self,
-        source: &Link,
-        directive_import: &Import,
-    ) -> Option<DirectiveCompositionSpecification> {
-        let specs = self.get_definition(&source.url)?.directive_specs();
-        let spec = specs
-            .iter()
-            .find(|s| *s.name() == directive_import.element)?;
-        let directive_spec: DirectiveSpecification = spec.as_any().downcast_ref().cloned()?;
-        directive_spec.composition
-    }
-}
-
-pub(crate) static SPEC_REGISTRY: LazyLock<SpecRegistry> = LazyLock::new(|| {
-    let mut registry = SpecRegistry::new();
-    registry.extend(&AUTHENTICATED_VERSIONS);
-    registry.extend(&CACHE_TAG_VERSIONS);
-    registry.extend(&CONNECT_VERSIONS);
-    registry.extend(&CONTEXT_VERSIONS);
-    registry.extend(&COST_VERSIONS);
-    registry.extend(&FEDERATION_VERSIONS);
-    registry.extend(&INACCESSIBLE_VERSIONS);
-    registry.extend(&POLICY_VERSIONS);
-    registry.extend(&REQUIRES_SCOPES_VERSIONS);
-    registry.extend(&TAG_VERSIONS);
-    registry.extend(&JOIN_VERSIONS);
-    registry
-});
