+++ apollo-federation/src/merge.rs
@@ -42,7 +42,6 @@ use itertools::Itertools;
 use crate::ValidFederationSubgraph;
 use crate::ValidFederationSubgraphs;
 use crate::error::FederationError;
-use crate::link::LinksMetadata;
 use crate::link::federation_spec_definition::FEDERATION_EXTERNAL_DIRECTIVE_NAME_IN_SPEC;
 use crate::link::federation_spec_definition::FEDERATION_FIELDS_ARGUMENT_NAME;
 use crate::link::federation_spec_definition::FEDERATION_FROM_ARGUMENT_NAME;
@@ -56,6 +55,7 @@ use crate::link::inaccessible_spec_definition::INACCESSIBLE_DIRECTIVE_NAME_IN_SP
 use crate::link::inaccessible_spec_definition::InaccessibleSpecDefinition;
 use crate::link::join_spec_definition::EnumValue;
 use crate::link::join_spec_definition::JOIN_OVERRIDE_LABEL_ARGUMENT_NAME;
+use crate::link::metadata::LinksMetadata;
 use crate::link::spec::Identity;
 use crate::link::spec::Version;
 use crate::link::spec_definition::SpecDefinition;
