+++ apollo-federation/src/schema/blueprint.rs
@@ -12,7 +12,6 @@ use crate::error::MultipleFederationErrors;
 use crate::error::SingleFederationError;
 use crate::error::suggestion::did_you_mean;
 use crate::error::suggestion::suggestion_list;
-use crate::link::DEFAULT_LINK_NAME;
 use crate::link::Link;
 use crate::link::federation_spec_definition::FED_1;
 use crate::link::federation_spec_definition::FEDERATION_FIELDS_ARGUMENT_NAME;
@@ -23,6 +22,7 @@ use crate::link::federation_spec_definition::FEDERATION_VERSIONS;
 use crate::link::federation_spec_definition::FederationSpecDefinition;
 use crate::link::federation_spec_definition::fed1_link_imports;
 use crate::link::federation_spec_definition::get_federation_spec_definition_from_subgraph;
+use crate::link::link_spec_definition::LINK_DIRECTIVE_NAME_IN_SPEC;
 use crate::link::link_spec_definition::LinkSpecDefinition;
 use crate::link::spec::Identity;
 use crate::link::spec::Url;
@@ -55,7 +55,7 @@ impl FederationBlueprint {
         schema: &mut FederationSchema,
         directive: &Component<Directive>,
     ) -> Result<Option<DirectiveDefinitionPosition>, FederationError> {
-        if directive.name != DEFAULT_LINK_NAME {
+        if directive.name != LINK_DIRECTIVE_NAME_IN_SPEC {
             return Ok(None);
         }
