+++ apollo-federation/src/schema/position.rs
@@ -35,7 +35,7 @@ use crate::error::CompositionError;
 use crate::error::FederationError;
 use crate::error::MultipleFederationErrors;
 use crate::error::SingleFederationError;
-use crate::link::database::links_metadata;
+use crate::link::metadata::LinksMetadata;
 use crate::link::spec_definition::SpecDefinition;
 use crate::merger::merge_argument::HasArguments;
 use crate::merger::merge_argument::HasDefaultValue;
@@ -2098,7 +2098,7 @@ impl SchemaDefinitionPosition {
             .directives
             .insert(index, directive);
         self.insert_directive_name_references(&mut schema.referencers, &name)?;
-        schema.links_metadata = links_metadata(&schema.schema)?.map(Box::new);
+        schema.links_metadata = LinksMetadata::from_schema(&schema.schema)?.map(Box::new);
         Ok(())
     }
 
@@ -2115,7 +2115,7 @@ impl SchemaDefinitionPosition {
             .directives
             .retain(|other_directive| other_directive.name != name);
         if is_link {
-            schema.links_metadata = links_metadata(&schema.schema)?.map(Box::new);
+            schema.links_metadata = LinksMetadata::from_schema(&schema.schema)?.map(Box::new);
         }
         Ok(())
     }
@@ -7977,7 +7977,7 @@ impl FederationSchema {
     }
 
     pub(crate) fn collect_links_metadata(&mut self) -> Result<(), FederationError> {
-        self.links_metadata = links_metadata(self.schema())?.map(Box::new);
+        self.links_metadata = LinksMetadata::from_schema(self.schema())?.map(Box::new);
         Ok(())
     }
