+++ apollo-router/src/spec/schema.rs
@@ -14,7 +14,7 @@ use apollo_federation::Supergraph;
 use apollo_federation::connectors::expand::Connectors;
 use apollo_federation::connectors::expand::ExpansionResult;
 use apollo_federation::connectors::expand::expand_connectors;
-use apollo_federation::link::database::links_metadata;
+use apollo_federation::link::metadata::LinksMetadata;
 use apollo_federation::link::spec::Identity;
 use apollo_federation::router_supported_supergraph_specs;
 use apollo_federation::schema::ValidFederationSchema;
@@ -315,7 +315,7 @@ impl Schema {
 
     /// This function assumes `@link` usage is valid in the schema, and will return `false` if not.
     pub(crate) fn has_spec(&self, spec_identity: &Identity, expected_version_range: &str) -> bool {
-        let Ok(Some(metadata)) = links_metadata(self.supergraph_schema()) else {
+        let Ok(Some(metadata)) = LinksMetadata::from_schema(self.supergraph_schema()) else {
             return false;
         };
         let Some(link) = metadata.for_identity(spec_identity) else {
@@ -338,7 +338,7 @@ impl Schema {
         expected_version_range: &str,
         default: &Name,
     ) -> Option<String> {
-        let metadata = links_metadata(schema).ok()??;
+        let metadata = LinksMetadata::from_schema(schema).ok()??;
         let link = metadata.for_identity(spec_identity)?;
         let semver_version = Version::parse(format!("{}.0", link.url.version).as_str()).ok()?;
         let version_range = VersionReq::parse(expected_version_range).ok()?;
